Removing everything that depends on time to work
